pyproject.toml
[tool.poetry]
name = "stunning-potato"
version = "0.1.0"
description = ""
authors = ["thearyadev <personal@aryankothari.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.30.0"
fastapi = "^0.95.1"
psycopg2 = "^2.9.6"
black = "^23.3.0"
isort = "^5.12.0"
rich = "^13.3.5"
beautifulsoup4 = "^4.12.2"
pillow = "^9.5.0"
beartype = "^0.14.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
